+++ openmls/src/group/mls_group/commit_builder.rs
@@ -201,7 +201,7 @@ impl<'a, T, G: BorrowMut<MlsGroup>> CommitBuilder<'a, T, G> {
 
 impl MlsGroup {
     /// Returns a builder for commits.
-    pub fn commit_builder(&mut self) -> CommitBuilder<Initial> {
+    pub fn commit_builder(&mut self) -> CommitBuilder<'_, Initial> {
         CommitBuilder::<'_, Initial, &mut MlsGroup>::new(self)
     }
 }
@@ -227,7 +227,7 @@ impl<'a> CommitBuilder<'a, Initial, &mut MlsGroup> {
         self.stage.own_proposals.extend(
             key_packages
                 .into_iter()
-                .map(|key_package| Proposal::Add(AddProposal { key_package })),
+                .map(|key_package| Proposal::add(AddProposal { key_package })),
         );
         self
     }
@@ -238,7 +238,7 @@ impl<'a> CommitBuilder<'a, Initial, &mut MlsGroup> {
         self.stage.own_proposals.extend(
             removed
                 .into_iter()
-                .map(|removed| Proposal::Remove(RemoveProposal { removed })),
+                .map(|removed| Proposal::remove(RemoveProposal { removed })),
         );
         self
     }
@@ -248,7 +248,7 @@ impl<'a> CommitBuilder<'a, Initial, &mut MlsGroup> {
     pub fn propose_group_context_extensions(mut self, extensions: Extensions) -> Self {
         self.stage
             .own_proposals
-            .push(Proposal::GroupContextExtensions(
+            .push(Proposal::group_context_extensions(
                 GroupContextExtensionProposal::new(extensions),
             ));
         self